Character list fragment gets characters from api

+++ app/src/main/java/com/example/hpcharactersapp/domain/model/Wand.kt
@@ -2,6 +2,6 @@ package com.example.hpcharactersapp.domain.model
 
 data class Wand(
     val core: String,
-    val length: Int,
+    val length: Double,
     val wood: String
 )
